backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt==3.2.0",
    "faker>=38.2.0",
    "fastapi>=0.121.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "psycopg[binary]>=3.2.13",
    "pydantic>=2.12.4",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
